+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- ReverseDsc
+  - Changed form Size to dynamic Width
+  - Change column width to calc /3 of Form.
+  - Export form is now more dynamic / responsive
+  
+
 ## [4.8.0] - 2021-08-31
 
 ### Added